Restructure project files.
- Move actual app files into src directory.
