+++ frontend/cypress/integration/signInForm.spec.js
@@ -1,6 +1,6 @@
 import faker from 'faker'
 
-import { SignInForm } from '@/models/SignInForm'
+import { SignInForm } from '../../src/models/SignInForm'
 
 describe('Log in form validation', () => {
   beforeEach(() => {
