+++ frontend/src/components/NavBar.module.scss
