+++ frontend/src/components/SignInForm.js
@@ -11,7 +11,7 @@ import {
 
 import OAuthButtons from './OAuthButtons'
 import useForm from '../hooks/useForm'
-import { SignInForm } from '@/models/SignInForm'
+import { SignInForm } from '@models/SignInForm'
 import { useRouter } from 'next/router'
 
 const endpoint = `${process.env.KITSPACE_GITEA_URL}/user/kitspace/sign_in`
