+++ frontend/src/components/SignUpForm.js
@@ -4,8 +4,8 @@ import { Button, Form, Header, Input, Message, Segment } from 'semantic-ui-react
 import { isEmpty } from 'lodash'
 
 import OAuthButtons from './OAuthButtons'
-import useForm from '../hooks/useForm'
-import { SignUpForm } from '@/models/SignUpForm'
+import useForm from '@hooks/useForm'
+import { SignUpForm } from '@models/SignUpForm'
 import { useRouter } from 'next/router'
 
 const endpoint = `${process.env.KITSPACE_GITEA_URL}/user/kitspace/sign_up`
@@ -76,8 +76,7 @@ export default function () {
         >
           Success!
         </Message.Header>
-        {errors.msg ||
-          apiResponse.message || 'Logging you in...'}
+        {errors.msg || apiResponse.message || 'Logging you in...'}
       </Message>
       <Form>
         <Segment>
