+++ frontend/src/pages/1-click-bom.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Link from 'next/link'
 
-import { Page } from '@/components/Page'
+import { Page } from '@components/Page'
 
 export default function OneClickBom() {
   return (
