+++ frontend/src/pages/bom-builder.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Page } from '@/components/Page'
+import { Page } from '@components/Page'
 
 export default function BomBuilder() {
   return (
