+++ frontend/src/pages/buy/cancelled.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { Page } from '@/components/Page'
+import { Page } from '@components/Page'
 import styles from './styles.module.scss'
 
 export default function CanceledPage() {
